requirements.txt
fastapi
uvicorn
scikit-learn
numpy
matplotlib
requests
wordcloud
paho-mqtt
mlflow
gensim
